refactor: split component tests by domain
1.component_test.go (1671 lines) → 6 focused files by domain
2.basic_test.go, reconnect_test.go, callback_test.go, lifecycle_test.go, misc_test.go
3.testhelpers_test.go: shared helpers (dialWithMock, echoLoop, fakeClock, etc.)
4.Remove TestComponent_ prefix — no longer needed without integration tests
