+++ testhelpers_test.go
@@ -1,0 +1,175 @@
+package client_test
+
+import (
+	"errors"
+	"net/http"
+	"sync"
+	"sync/atomic"
+	"testing"
+	"time"
+
+	wspulse "github.com/wspulse/core"
+
+	"github.com/wspulse/client-go"
+)
+
+// ── fakeClock ────────────────────────────────────────────────────────────────
+
+// fakeClock replaces both NewTimer (backoff) and NewTicker (heartbeat) with
+// controllable fakes. No real timers fire — tests drive time explicitly.
+type fakeClock struct {
+	mu      sync.Mutex
+	timers  []*fakeTimerEntry
+	tickers []*fakeTickerEntry
+}
+
+type fakeTimerEntry struct {
+	d     time.Duration
+	timer *time.Timer
+}
+
+type fakeTickerEntry struct {
+	d      time.Duration
+	ticker *time.Ticker
+}
+
+func newFakeClock() *fakeClock { return &fakeClock{} }
+
+// NewTimer returns a stopped timer that will not fire on its own.
+func (fc *fakeClock) NewTimer(d time.Duration) *time.Timer {
+	t := time.NewTimer(time.Hour)
+	t.Stop()
+	fc.mu.Lock()
+	fc.timers = append(fc.timers, &fakeTimerEntry{d: d, timer: t})
+	fc.mu.Unlock()
+	return t
+}
+
+// NewTicker returns a stopped ticker that will never fire on its own.
+func (fc *fakeClock) NewTicker(d time.Duration) *time.Ticker {
+	t := time.NewTicker(time.Hour)
+	t.Stop()
+	fc.mu.Lock()
+	fc.tickers = append(fc.tickers, &fakeTickerEntry{d: d, ticker: t})
+	fc.mu.Unlock()
+	return t
+}
+
+// TimerCount returns the number of registered timers.
+func (fc *fakeClock) TimerCount() int {
+	fc.mu.Lock()
+	defer fc.mu.Unlock()
+	return len(fc.timers)
+}
+
+// TickerCount returns the number of registered tickers.
+func (fc *fakeClock) TickerCount() int {
+	fc.mu.Lock()
+	defer fc.mu.Unlock()
+	return len(fc.tickers)
+}
+
+// ── helpers ──────────────────────────────────────────────────────────────────
+
+// echoLoop reads from mt.writeCh and injects the written data back as a
+// text message, simulating an echo server. Stops when done is closed.
+func echoLoop(mt *mockTransport, done chan struct{}) {
+	for {
+		select {
+		case w := <-mt.writeCh:
+			// Only echo text messages (type 1), skip pings (9) and close (8).
+			if w.messageType == 1 {
+				mt.InjectMessage(1, w.data)
+			}
+		case <-done:
+			return
+		}
+	}
+}
+
+// fireBackoffTimers spins up a goroutine that fires up to maxTimers backoff
+// timers on fc as they are created. Returns a stop function that must be called
+// to prevent goroutine leaks (typically via defer).
+func fireBackoffTimers(fc *fakeClock, maxTimers int) (stop func()) {
+	stopCh := make(chan struct{})
+	stopped := make(chan struct{})
+	go func() {
+		defer close(stopped)
+		for i := 0; i < maxTimers; i++ {
+			deadline := time.Now().Add(time.Second)
+			for {
+				select {
+				case <-stopCh:
+					return
+				default:
+				}
+				fc.mu.Lock()
+				count := len(fc.timers)
+				fc.mu.Unlock()
+				if count > i {
+					fc.mu.Lock()
+					fc.timers[count-1].timer.Reset(0)
+					fc.mu.Unlock()
+					break
+				}
+				if time.Now().After(deadline) {
+					break
+				}
+				time.Sleep(time.Millisecond)
+			}
+		}
+	}()
+	return func() {
+		close(stopCh)
+		<-stopped
+	}
+}
+
+// dialWithMock creates a mock dialer with a single transport and dials using it.
+// Returns the client, mock transport, and fake clock.
+func dialWithMock(t *testing.T, opts ...client.ClientOption) (client.Client, *mockTransport, *fakeClock) {
+	t.Helper()
+	mt := newMockTransport()
+	fc := newFakeClock()
+	md := newMockDialer(mockDialResult{transport: mt})
+	allOpts := []client.ClientOption{
+		client.WithDialer(md),
+		client.WithClock(fc),
+	}
+	allOpts = append(allOpts, opts...)
+	c, err := client.Dial("ws://mock", allOpts...)
+	if err != nil {
+		t.Fatalf("Dial failed: %v", err)
+	}
+	return c, mt, fc
+}
+
+// headerCapturingDialer is a mock dialer that captures the headers passed to Dial.
+type headerCapturingDialer struct {
+	transport *mockTransport
+	onDial    func(http.Header)
+	called    atomic.Bool
+}
+
+func (d *headerCapturingDialer) Dial(_ string, header http.Header) (wspulse.Transport, error) {
+	if d.onDial != nil {
+		d.onDial(header)
+	}
+	if d.called.CompareAndSwap(false, true) {
+		return d.transport, nil
+	}
+	return nil, errors.New("headerCapturingDialer: no more transports")
+}
+
+// failEncodeCodecComponent is a test codec whose Encode always returns an error.
+type failEncodeCodecComponent struct{}
+
+func (failEncodeCodecComponent) Encode(wspulse.Frame) ([]byte, error) {
+	return nil, errors.New("wspulse: encode fail")
+}
+
+func (failEncodeCodecComponent) Decode(data []byte) (wspulse.Frame, error) {
+	return wspulse.Frame{}, nil
+}
+
+func (failEncodeCodecComponent) FrameType() int { return 1 }
